Android methods compatable with testcases
