+++ CBLClient/Apps/JavaRequestHandlers/src/main/java/com/couchbase/mobiletestkit/javacommon/RequestHandler/DatabaseRequestHandler.java
@@ -387,75 +387,6 @@ public class DatabaseRequestHandler {
 
         return data;
     }
-
-    public Scope defaultScope(Args args) throws CouchbaseLiteException {
-        Database db = args.get("database");
-        return db.getDefaultScope();
-    }
-
-    public Collection defaultCollection(Args args) throws CouchbaseLiteException {
-        Database db =args.get("database");
-        return db.getDefaultCollection();
-
-    }
-
-    public Collection createCollection(Args args) throws CouchbaseLiteException {
-        Database db = args.get("database");
-        String collectionName = args.get("collectionName");
-        String scopeName = args.get("scopeName");
-        if (scopeName == null) {
-            return db.createCollection(collectionName);
-        }
-        else {
-            return db.createCollection(collectionName, scopeName);
-        }
-    }
-
-    public void deleteCollection(Args args) throws CouchbaseLiteException {
-        Database db = args.get("database");
-        String collectionName = args.get("collectionName");
-        String scopeName = args.get("scopeName");
-        if (scopeName == null) {
-            db.deleteCollection(collectionName);
-            return;
-        }
-        else {
-            db.deleteCollection(collectionName, scopeName);
-            return;
-        }
-    }
-
-    public static Set<Collection> collectionsInScope(Database db, String scopeName) throws CouchbaseLiteException {
-        if (scopeName == null) {
-            return db.getCollections();
-        }
-        else {
-            return db.getCollections(scopeName);
-        }
-    }
-
-    public ArrayList<String> collectionNames (Args args) throws CouchbaseLiteException {
-        Database db = args.get("database");
-        String scopeName = args.get("scopeName");
-        ArrayList<String> collectionsNames = new ArrayList<>();
-        Set<Collection> collections = collectionsInScope(db, scopeName);
-        for (Collection collectionObject: collections) {
-            collectionsNames.add(collectionObject.getName());
-        }
-        return collectionsNames;
-    }
-
-    public Collection collectionObject(Args args) throws CouchbaseLiteException {
-        Database db = args.get("database");
-        String collectionName = args.get("collectionName");
-        String scopeName = args.get("scopeName");
-        if (scopeName == null) {
-            return db.getCollection(collectionName);
-        }
-        else {
-            return db.getCollection(collectionName, scopeName);
-        }
-    }
 }
 
 class MyDatabaseChangeListener implements DatabaseChangeListener {
